riscv_qemu: Initalize PLIC interrupt priorities.
When using default Qemu 7.x build PLIC interrupts do not seem to work.
I'm not sure if this is caused by Qemu itself, or newer OpenSBI version.
Still, the actual root cause is pretty simple. Genode does not try to
initialize PLIC priority threshold and interrupt priorities. As result
the threshold is by default set to value 7, while individual irq
priorities are 0. IRQ priority of 0 means that the interrupt is disabled
according to PLIC 1.0-rc2 specification.

Fix this by setting interrupt threshould to 0 and interrupt priorities
to 1.

issue #2